app/core/tests/test_admin.py
"""
Tests for Django admin modifications.
"""
from django.test import Client, TestCase
from django.urls import reverse
from django.contrib.auth import get_user_model


class AdminSiteTests(TestCase):
    """Tests for Django admin."""

    def setUp(self):
        """Create user and client."""
        self.client = Client()
        self.admin_user = get_user_model().objects.create_superuser(
            email="admin@example.com", password="testpass123"
        )
        self.client.force_login(self.admin_user)
        self.user = get_user_model().objects.create_user(
            email="user@example.com", password="testpass123", name="Test User"
        )

    def test_users_list(self):
        """Test that users are listed on page."""
        # `reverse()` fetches the URL where the changelist is visible.
        url = reverse("admin:core_user_changelist")
        res = self.client.get(url)

        self.assertContains(res, self.user.name)
        self.assertContains(res, self.user.email)

    def test_edit_user_page(self):
        """Test that the edit user page works."""
        # Retrieves URL with format
        # "http://localhost:8000/admin/core/user/<user_id>/change/"
        url = reverse("admin:core_user_change", args=[self.user.id])
        res = self.client.get(url)

        # We just check if the modify user page is accessible.
        self.assertEqual(res.status_code, 200)

    def test_create_user_page(self):
        """Test the create user page works."""
        url = reverse("admin:core_user_add")
        res = self.client.get(url)

        self.assertEqual(res.status_code, 200)
